requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
openai-agents-mcp
python-dotenv